feat: add to ocean tag status
